requirements.txt
Flask==3.0.0
requests==2.31.0
scikit-learn==1.3.1
pandas==2.1.1
numpy==1.26.0
python-dotenv==1.2.1
